move watchers used only in Arboria effect
