--- Mage/src/main/java/mage/watchers/common/PermanentsEnteredBattlefieldYourLastTurnWatcher.java
@@ -1,65 +1,0 @@
-package mage.watchers.common;
-
-import mage.constants.WatcherScope;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
-import mage.watchers.Watcher;
-
-import java.util.*;
-
-/**
- * @author LevelX2 (spjspj)
- */
-public class PermanentsEnteredBattlefieldYourLastTurnWatcher extends Watcher {
-
-    private final Map<UUID, List<Permanent>> enteringBattlefield = new HashMap<>();
-    private final Map<UUID, List<Permanent>> enteringBattlefieldLastTurn = new HashMap<>();
-    private UUID lastActivePlayer = null;
-
-    public PermanentsEnteredBattlefieldYourLastTurnWatcher() {
-        super(WatcherScope.GAME);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        lastActivePlayer = game.getActivePlayerId();
-
-        if (event.getType() == GameEvent.EventType.ENTERS_THE_BATTLEFIELD) {
-            Permanent perm = game.getPermanentEntering(event.getTargetId());
-            if (perm == null) {
-                perm = game.getPermanent(event.getTargetId());
-            }
-            if (perm != null) {
-                List<Permanent> permanents;
-                if (!enteringBattlefield.containsKey(perm.getControllerId())) {
-                    permanents = new ArrayList<>();
-                    enteringBattlefield.put(perm.getControllerId(), permanents);
-                } else {
-                    permanents = enteringBattlefield.get(perm.getControllerId());
-                }
-                permanents.add(perm.copy()); // copy needed because attributes like color could be changed later
-            }
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        if (enteringBattlefieldLastTurn != null
-                && lastActivePlayer != null
-                && enteringBattlefieldLastTurn.get(lastActivePlayer) != null) {
-            enteringBattlefieldLastTurn.remove(lastActivePlayer);
-        }
-        enteringBattlefieldLastTurn.putAll(enteringBattlefield);
-        enteringBattlefield.clear();
-        lastActivePlayer = null;
-    }
-
-    public List<Permanent> getPermanentsEnteringOnPlayersLastTurn(Game game, UUID playerId) {
-        if (game.isActivePlayer(playerId)) {
-            return enteringBattlefield.get(playerId);
-        }
-        return enteringBattlefieldLastTurn.get(playerId);
-    }
-}
